refactor: clean up testing files
